Save tracker time and cpu time everytime sample() is called
This is the lowest level to get those times so this should be the
perfect place to get them and write them to a file. We did this to check
how big the lag between producing data in the eyetracker and getting
them from PyGaze is, and whether it's a constant lag or building up over
time.